Trait command_fds::CommandFdExt

pub trait CommandFdExt {
    fn fd_mappings(
        &mut self, 
        mappings: Vec<FdMapping>
    ) -> Result<&mut Self, FdMappingCollision>;
    fn preserved_fds(&mut self, fds: Vec<RawFd>) -> &mut Self;
}

Extension to add file descriptor mappings to a Command.

Required methods

fn fd_mappings(
&mut self,
mappings: Vec<FdMapping>
) -> Result<&mut Self, FdMappingCollision>

Adds the given set of file descriptors to the command.

Warning: Calling this more than once on the same command, or attempting to run the same command more than once after calling this, may result in unexpected behaviour.

fn preserved_fds(&mut self, fds: Vec<RawFd>) -> &mut Self

Adds the given set of file descriptors to be passed on to the child process when the command is run.