Crate command_fds

A library for passing arbitrary file descriptors when spawning child processes.

Example

use command_fds::{CommandFdExt, FdMapping};
use std::fs::File;
use std::os::unix::io::AsRawFd;
use std::process::Command;

// Open a file.
let file = File::open("Cargo.toml").unwrap();

// Prepare to run `ls -l /proc/self/fd` with some FDs mapped.
let mut command = Command::new("ls");
command.arg("-l").arg("/proc/self/fd");
command
    .fd_mappings(vec![
        // Map `file` as FD 3 in the child process.
        FdMapping {
            parent_fd: file.as_raw_fd(),
            child_fd: 3,
        },
        // Map this process's stdin as FD 5 in the child process.
        FdMapping {
            parent_fd: 0,
            child_fd: 5,
        },
    ])
    .unwrap();

// Spawn the child process.
let mut child = command.spawn().unwrap();
child.wait().unwrap();

Structs

FdMapping

A mapping from a file descriptor in the parent to a file descriptor in the child, to be applied when spawning a child process.

FdMappingCollision

Error setting up FD mappings, because there were two or more mappings for the same child FD.

Traits

CommandFdExt

Extension to add file descriptor mappings to a Command.